Refactor: ProcessConfirm 컴포넌트 분리, 전역 알림 모달 생성
