+++ src/renderer/Components/SignUp/index.jsx
@@ -1,6 +1,6 @@
+import axios from "axios";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import axios from "axios";
 import { getAuth, createUserWithEmailAndPassword } from "firebase/auth";
 
 import Modal from "../Modal";
